Made StagePlanner its own component
